feat: add jsc JSONTestSuite
Since the existing test suite didn't catch any of the reported problems
with the JSON parsing, adding another larger suite that's used to
compare popular JSON parsers should!

Obviously because this is so large it can be trimmed, but the hope is to
find overlap between the feedback and the test-cases here.