feat: use a single random id for account id
Instead of having a known salt which is used to derive the create2
salt, and a static id which is used to derive an account id, use
the same hashed id for both.

This avoids the problem of having contention over account ids, but
removes the ability for a deployer to select a specific account id.
(an account mapping of msg.sender to account id would have the
same effect)