feat: remove standalone precompile test
Even including this as an internal function was confusing for
external reviewers. Now that there are other tests to verify the
full message flow, moving this to a test-only module wouldn't make
much sense. If we were going to add this back it would be to expose
the combination of _createMessage and callVerifier so the test would
actually cover a part of the webauthn module flow instead of just the
verifier caller.